merged-dave-eddie-changes
