+++ client/package.json
@@ -29,7 +29,7 @@
         "mui": "^0.0.1",
         "path": "^0.12.7",
         "react": "^18.2.0",
-        "react-achievements": "^1.2.0",
+        "react-achievements": "^1.3.1",
         "react-bootstrap": "^2.10.2",
         "react-burger-menu": "^3.0.9",
         "react-confetti": "^6.1.0",
