+++ client/src/components/JoyRideTour.jsx
@@ -1,49 +1,49 @@
 import React, { useContext } from 'react';
 import Joyride, { STATUS } from 'react-joyride';
-import { AppContext } from '../context/AppContext';
-
+import { AppContext } from '../context/AppContext.jsx';
 
 const steps = [
   {
-    target: '.landing-header',
-    content: 'Welcome to SimpliTrac! This is a simple to use financial tracker that allows you to record expenses on the go.',
+    target: 'body',
+    content: 'Welcome to SimpliTrac! This is a simple-to-use financial tracker that allows you to record expenses on the go.',
+    placement: 'center',
     disableBeacon: true,
   },
   {
-    target: '.buttons-container button:nth-child(1)',
-    content: 'With our intuitive AI-driven camera setup, you can snap pictures of your receipts on the fly and record your daily expenses.',
+    target: '.chakra-container',
+    content: 'This is your main dashboard where you can view and manage your expenses.',
   },
   {
-    target: '.landing-content',
-    content: 'You can manually enter your expenses on the main page with our easy to use Expense Table.',
+    target: '.expenses-table',
+    content: 'Here you can manually enter your expenses or view your recent transactions.',
   },
   {
-    target: '.buttons-container button:nth-child(2)',
-    content: 'See where your money is going in real time with our handy Chart feature.',
+    target: '[data-tour="scan-receipt"]',
+    content: 'Use our intuitive AI-driven camera to snap pictures of your receipts and record expenses instantly.',
   },
   {
-    target: '.buttons-container button:nth-child(3)',
-    content: 'Use the Edit Transaction function to correct any errors in your expense processing.',
+    target: '[data-tour="edit-transactions"]',
+    content: 'Easily modify or correct any of your recorded expenses here.',
   },
   {
-    target: 'body',
-    content: 'Be sure to check your achievements tab to see how you rank among other users and earn badges while maintaining financial peace of mind!',
-    placement: 'center',
+    target: '[data-tour="expense-chart"]',
+    content: 'View a detailed breakdown of your spending habits with our interactive chart feature.',
   },
 ];
+const JoyrideTour = () => {
+ 
+  const { runTour, setRunTour } = useContext(AppContext)
 
-const JoyrideTour = ({ run, setRun }) => {
   const handleJoyrideCallback = (data) => {
     const { status } = data;
     if ([STATUS.FINISHED, STATUS.SKIPPED].includes(status)) {
-      setRun(false);
+      setRunTour(false);
     }
   };
-
   return (
     <Joyride
       steps={steps}
-      run={run}
+      run={runTour}
       continuous={true}
       showSkipButton={true}
       showProgress={true}
@@ -56,5 +56,4 @@ const JoyrideTour = ({ run, setRun }) => {
     />
   );
 };
-
 export default JoyrideTour;
