+++ client/src/context/AppContext.jsx
@@ -9,7 +9,7 @@ const AppContext = createContext({});
 
 const AppProvider = ({ children }) => {
 
-    const detectDevice = () => {
+    const detectDevice = () => { 
         if (isDesktop) return 'desktop'
 
         return 'mobile'
@@ -21,7 +21,7 @@ const AppProvider = ({ children }) => {
         }
         setScreen(screen);
     };
-
+ 
     const toggleCategoriesList = () => {
         setShowCategories(!showCategories);
     };
@@ -38,7 +38,10 @@ const AppProvider = ({ children }) => {
     const [formData, setFormData] = useState(new FormData());
     const [isUpdating, setIsUpdating] = useState(false)
     const [showHamburger, setShowHamburger] = useState(false)
+    const [showCategories, setShowCategories] = useState(false)
     const [categoriesSelected, setCategoriesSelected] = useState(false);
+    const [runTour, setRunTour] = useState(false);
+    const [runEditTransactionsTour, setRunEditTransactionsTour] =useState(false);
 
 
 
@@ -87,6 +90,9 @@ const AppProvider = ({ children }) => {
         setFormData(new FormData());
         setIsUpdating(false)
         setShowHamburger(false)
+        setShowCategories(false)
+        setRunTour(false)
+        setRunEditTransactionsTour(false)
     };
 
     // updating data based on user state
@@ -109,9 +115,12 @@ const AppProvider = ({ children }) => {
         resetAppState,
         isUpdating, setIsUpdating,
         showHamburger, setShowHamburger,
+        runTour, setRunTour,
+        showCategories, setShowCategories,
         renderNewScreen,
         toggleCategoriesList,
         categoriesSelected, setCategoriesSelected,
+        runEditTransactionsTour, setRunEditTransactionsTour,
     };
