--- src/App.css
@@ -1,182 +1,0 @@
-:root {
-  --primary-color: #007BFF;
-  --primary-hover: #0056b3;
-  --text-color: #495057;
-  --background-color: #f3f4f5;
-  --border-color: #ced4da;
-  --error-color: #dc3545;
-  --error-hover: #c82333;
-  --chakra-colors-primary-main: #1B263B
-}
-
-/* Base styles */
-body, #root {
-  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
-  line-height: 1.5;
-  color: var(--text-color);
-  background-color: var(--background-color);
-  margin: 0;
-  min-width: 320px;
-  min-height: 100vh;
-}
-
-#root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
-}
-
-/* Typography */
-h1 {
-  font-size: 2.5em;
-  line-height: 1.1;
-  color: var(--text-color);
-}
-
-a {
-  color: var(--primary-color);
-  text-decoration: none;
-}
-
-a:hover {
-  color: var(--primary-hover);
-}
-
-/* Buttons */
-button, .custom-button {
-  padding: 0.75rem;
-  background-color: var(--primary-color);
-  border: none;
-  border-radius: 5px;
-  color: white;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-button:hover, .custom-button:hover {
-  background-color: var(--primary-hover);
-}
-
-/* buttons on edit page */
-.edit-buttons {
-  display: flex;
-  justify-content: center;
-  align-items: center; 
-  gap: 5px; 
-}
-
-.edit-left-button {
-  display: flex;
-  gap: 16px; /* Adds space between Cancel and Back buttons */
-}
-
-.edit-right-button {
-  margin-left: auto;
-}
-
-/* space between columns/rows in table */
-.edit-table {
-  width: 100%;
-  border-collapse: separate;
-  border-spacing: 3px; 
-}
-
-/* Column headers */
-.edit-table th {
-  font-weight: bold;
-  font-size: 18px; 
-}
-
-/* Forms */
-form {
-  width: 100%;
-  padding: 2rem;
-  background: white;
-  border-radius: 10px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-form div {
-  margin-bottom: 1rem;
-  text-align: left;
-}
-
-form label {
-  display: block;
-  margin-bottom: 0.5rem;
-  font-weight: bold;
-}
-
-form input, form select {
-  width: 100%;
-  padding: 0.75rem;
-  border: 1px solid var(--border-color);
-  border-radius: 5px;
-  font-size: 1rem;
-}
-
-form input:focus, form select:focus {
-  border-color: var(--primary-color);
-  box-shadow: 0 0 0 0.2rem rgba(0, 123, 255, 0.25);
-}
-
-/* Modal Styles */
-.modal-overlay {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: rgba(0, 0, 0, 0.75);
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-}
-
-.modal-content {
-  width: 90%;
-  max-width: 500px;
-  padding: 20px;
-  background: white;
-  border-radius: 8px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-/* List Styles */
-.source-type-modal__list {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-}
-
-.source-type-modal__list-item {
-  margin-bottom: 10px;
-}
-
-/* Checkbox Styles */
-.source-type-modal__list-item input[type="checkbox"] {
-  width: 20px;
-  height: 20px;
-  margin-right: 8px;
-}
-
-/* Delete Button */
-.delete-button {
-  background-color: var(--error-color);
-}
-
-.delete-button:hover {
-  background-color: var(--error-hover);
-}
-.transaction-edit-row {
-  display: flex;
-  align-items: center;
-  margin-bottom: 10px;
-} 
-
-.transaction-edit-row input,
-.transaction-edit-row select {
-  margin-right: 10px;
-}
